animation and folder restructure
