+++ src/components/sections/ExpertiseSection.js
@@ -1,0 +1,18 @@
+import React from "react";
+import Markdown from 'markdown-to-jsx';
+import Reveal from 'react-reveal/Reveal';
+
+export default function ExpertiseSection(props) {
+  const {
+    description,
+    techstacks
+  } = props
+
+  return (
+    <React.Fragment>
+      <section className="expertise">
+        
+      </section>
+    </React.Fragment>
+  );
+}
